Update dependencies and enhance cashflow examples with new parameters
### Summary
- Added `num-traits` dependency to `Cargo.toml` for improved numerical operations.
- Enhanced various cashflow examples by introducing new parameters such as `rate`, `spread_bp`, and `floor_bp` to better illustrate cashflow calculations.
- Updated test cases to reflect changes in cashflow specifications and ensure accuracy in calculations.

### Changes
- **Cargo.toml**: Added `num-traits` dependency.
- **examples/portfolio/portfolio_example.rs**: Introduced `spread_bp` parameter.
- **examples/valuations/bond_custom_cashflows_example.rs**: Added `rate` parameter.
- **examples/valuations/enhanced_builders_example.rs**: Included `spread_bp` parameter.
- **examples/valuations/tranche_valuation_example.rs**: Added `floor_bp` parameter.
- **examples/statements/statements_phase6_example.rs**: Updated to include new rate calculations.
- **tests/**: Adjusted tests to validate new parameters in cashflow examples.

+++ finstack/valuations/src/cashflow/builder/compiler.rs
@@ -24,6 +24,7 @@ use crate::cashflow::builder::{AmortizationSpec, Notional};
 use finstack_core::dates::{BusinessDayConvention, Date, DayCount, StubKind, Tenor};
 use finstack_core::error::InputError;
 use finstack_core::money::Money;
+use rust_decimal::Decimal;
 
 use super::date_generation::{build_dates, build_dates_checked};
 use super::specs::{
@@ -79,14 +80,14 @@ pub(super) type FloatSchedule = (
 ///
 /// Fields:
 /// - `base` (`FeeBase`): Notional/cash‑based fee base.
-/// - `bps` (`f64`): Annualized basis points applied to the base.
+/// - `bps` (`Decimal`): Annualized basis points applied to the base. Uses Decimal for exact representation.
 /// - `dc` (`DayCount`): Day‑count convention for accrual.
 /// - `dates` (`Vec<Date>`): Inclusive/exclusive boundary dates for accrual periods.
 /// - `prev` (`HashMap<Date, Date>`): Mapping of each period end to its start.
 #[derive(Debug, Clone)]
 pub(super) struct PeriodicFee {
     pub(super) base: FeeBase,
-    pub(super) bps: f64,
+    pub(super) bps: Decimal,
     pub(super) dc: DayCount,
     pub(super) dates: Vec<Date>,
     pub(super) prev: finstack_core::collections::HashMap<Date, Date>,
@@ -191,12 +192,12 @@ pub(super) struct DateWindow {
 #[derive(Debug, Clone)]
 pub(super) enum CouponSpec {
     Fixed {
-        rate: f64,
+        rate: Decimal,
     },
     Float {
         index_id: finstack_core::types::CurveId,
-        margin_bp: f64,
-        gearing: f64,
+        margin_bp: Decimal,
+        gearing: Decimal,
         reset_lag_days: i32,
     },
 }
